+++ packages/yucca-sdk/orchestration-api/src/index.ts
@@ -1,5 +1,6 @@
 export * from './const';
 export { EventsGateway, type GatewayEvent } from './events/events.gateway';
+export type { ImmichDatabaseDumpConfig } from './moduleConfig';
 export * from './orchestrationApi.module';
 export { LoggingRepository } from './repositories/logging.repository';
 export { ModuleConfigRepository } from './repositories/moduleConfig.repository';