+++ packages/yucca-sdk/orchestration-ui/src/lib/fetch-client.ts
@@ -161,12 +161,18 @@ export type RunDto = {
     status: RunStatus;
     "type": RunType;
 };
+export type ImmichDatabaseDumpConfigDto = {
+    enabled: boolean;
+    keepLastAmount: number;
+};
 export type ImmichBackupStatusDto = {
     integration?: ImmichIntegrationDto;
     repository?: LocalRepositoryDto;
     backend?: BackendDto;
     schedule?: ScheduleDto;
     latestBackupRun?: RunDto;
+    databaseDump?: ImmichDatabaseDumpConfigDto;
+    databaseDumpWarningIgnored?: boolean;
 };
 export type ConfigureImmichIntegrationRequestDto = {
     name: string;
@@ -181,6 +187,10 @@ export type ConfigureImmichIntegrationRequestDto = {
 };
 export type ConfigureImmichIntegrationResponseDto = {
     repositoryId: string;
+};
+export type ConfigureImmichDatabaseDumpRequestDto = {
+    enabled?: boolean;
+    keepLastAmount?: number;
 };
 export type ImmichRollbackRequestDto = {
     repositoryId: string;
@@ -433,6 +443,19 @@ export function configureImmichIntegration(configureImmichIntegrationRequestDto:
         ...opts,
         method: "POST",
         body: configureImmichIntegrationRequestDto
+    })));
+}
+export function ignoreImmichDatabaseDumpWarning(opts?: Oazapfts.RequestOpts) {
+    return oazapfts.ok(oazapfts.fetchText("/api/yucca/integrations/immich/database-dump/ignore-warning", {
+        ...opts,
+        method: "POST"
+    }));
+}
+export function configureImmichDatabaseDump(configureImmichDatabaseDumpRequestDto: ConfigureImmichDatabaseDumpRequestDto, opts?: Oazapfts.RequestOpts) {
+    return oazapfts.ok(oazapfts.fetchText("/api/yucca/integrations/immich/database-dump", oazapfts.json({
+        ...opts,
+        method: "POST",
+        body: configureImmichDatabaseDumpRequestDto
     })));
 }
 export function startImmichRollback(immichRollbackRequestDto: ImmichRollbackRequestDto, opts?: Oazapfts.RequestOpts) {